# endpoint: Reduce CPU and memory usage of function moveNewFilesTo