This commit makes changes to the function moveNewFilesTo to reduce its complexity from O(n^2) to O(n) and decrease its CPU and memory usage. The function moveNewFilesTo is a utility function that may be called in the endpoint regeneration hot path. This commit makes the following changes to speed it up and reduce memory allocations:

* Use a WorkGroup to parallelize rename operations. Before this commit, these operations were done sequentially.
* Use a map to lookup if a file from the source directory exists in the destination directory. Before this commit, this lookup was done using a nested for loop, looping over slices that contain files in the source and destination directories.
* Use `os.File.Readdirnames` instead of os.ReadDir. The function `os.ReadDir` sorts the slice of returned files, which is not needed for this use case. Additionally, Readdirnames returns a string slice of file names, rather than a slice of `os.DirEntry` structs, meaning the name of each file does need to be accessed using an accessor method.

The benefits of these changes will be stronger as the number of files to move between two directories increases, however these changes also show significant benefits at a smaller scale (see the benchmarks below). In the future, more work can be done to look at the higher-level context of how an endpoint's eBPF programs are regenerated, to find greater optimizations.

This commit additionally introduces benchmark tests. Here are results comparing the implemention of moveNewFilesTo for before and after this commit using benchstat. Each benchmark was executed 10 times.

```
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/endpoint
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                    │ MoveNewFilesTo10.txt │    MoveNewFilesTo10-parallel.txt    │
                    │        sec/op        │   sec/op     vs base                │
MoveNewFilesTo10-20            9.227µ ± 1%   7.817µ ± 2%  -15.28% (p=0.000 n=10)

                    │ MoveNewFilesTo10.txt │    MoveNewFilesTo10-parallel.txt     │
                    │         B/op         │     B/op      vs base                │
MoveNewFilesTo10-20           1.992Ki ± 0%   1.338Ki ± 0%  -32.82% (p=0.000 n=10)

                    │ MoveNewFilesTo10.txt │   MoveNewFilesTo10-parallel.txt    │
                    │      allocs/op       │ allocs/op   vs base                │
MoveNewFilesTo10-20             48.00 ± 0%   31.00 ± 0%  -35.42% (p=0.000 n=10)
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/endpoint
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                    │ MoveNewFilesTo20.txt │    MoveNewFilesTo20-parallel.txt    │
                    │        sec/op        │   sec/op     vs base                │
MoveNewFilesTo20-20            12.07µ ± 1%   10.09µ ± 1%  -16.34% (p=0.000 n=10)

                    │ MoveNewFilesTo20.txt │    MoveNewFilesTo20-parallel.txt     │
                    │         B/op         │     B/op      vs base                │
MoveNewFilesTo20-20           3.450Ki ± 0%   2.380Ki ± 0%  -31.02% (p=0.000 n=10)

                    │ MoveNewFilesTo20.txt │   MoveNewFilesTo20-parallel.txt    │
                    │      allocs/op       │ allocs/op   vs base                │
MoveNewFilesTo20-20             74.00 ± 0%   45.00 ± 0%  -39.19% (p=0.000 n=10)
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/endpoint
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                    │ MoveNewFilesTo50.txt │    MoveNewFilesTo50-parallel.txt    │
                    │        sec/op        │   sec/op     vs base                │
MoveNewFilesTo50-20            23.25µ ± 1%   17.04µ ± 1%  -26.71% (p=0.000 n=10)

                    │ MoveNewFilesTo50.txt │    MoveNewFilesTo50-parallel.txt     │
                    │         B/op         │     B/op      vs base                │
MoveNewFilesTo50-20           7.384Ki ± 0%   4.745Ki ± 0%  -35.74% (p=0.000 n=10)

                    │ MoveNewFilesTo50.txt │   MoveNewFilesTo50-parallel.txt    │
                    │      allocs/op       │ allocs/op   vs base                │
MoveNewFilesTo50-20            152.00 ± 0%   86.00 ± 0%  -43.42% (p=0.000 n=10)
goos: linux
goarch: amd64
pkg: github.com/cilium/cilium/pkg/endpoint
cpu: 12th Gen Intel(R) Core(TM) i7-12700H
                     │ MoveNewFilesTo100.txt │   MoveNewFilesTo100-parallel.txt    │
                     │        sec/op         │   sec/op     vs base                │
MoveNewFilesTo100-20             45.55µ ± 2%   29.59µ ± 3%  -35.04% (p=0.000 n=10)

                     │ MoveNewFilesTo100.txt │    MoveNewFilesTo100-parallel.txt    │
                     │         B/op          │     B/op      vs base                │
MoveNewFilesTo100-20           14.251Ki ± 0%   9.204Ki ± 0%  -35.41% (p=0.000 n=10)

                     │ MoveNewFilesTo100.txt │   MoveNewFilesTo100-parallel.txt   │
                     │       allocs/op       │ allocs/op   vs base                │
MoveNewFilesTo100-20              278.0 ± 0%   151.0 ± 0%  -45.68% (p=0.000 n=10)
```

Please ensure your pull request adheres to the following guidelines:

- [ ] For first time contributors, read [Submitting a pull request](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#submitting-a-pull-request)
- [ ] All code is covered by unit and/or runtime tests where feasible.
- [ ] All commits contain a well written commit description including a title,
      description and a `Fixes: #XXX` line if the commit addresses a particular
      GitHub issue.
- [ ] If your commit description contains a `Fixes: <commit-id>` tag, then
      please add the commit author[s] as reviewer[s] to this issue.
- [ ] All commits are signed off. See the section [Developer’s Certificate of Origin](https://docs.cilium.io/en/stable/contributing/development/contributing_guide/#dev-coo)
- [ ] Provide a title or release-note blurb suitable for the release notes.
- [ ] Are you a user of Cilium? Please add yourself to the [Users doc](https://github.com/cilium/cilium/blob/main/USERS.md)
- [ ] Thanks for contributing!

<!-- Description of change -->

Fixes: #issue-number

```release-note
<!-- Enter the release note text here if needed or remove this section! -->
```