@learnitall 2023-06-16
Looks like we hit two flakes in the last test:

* K8sUpstreamNetConformance is full of `putEndpointIdTooManyRequests`, which is tracked in #24361. (Failed job: https://github.com/cilium/cilium/actions/runs/5284212566/jobs/9561415454?pr=26238)
* ClusterMesh failed on a pod-to-world test that curls google, which is tracked in #25442. (Failed job: https://github.com/cilium/cilium/actions/runs/5284646478/jobs/9562294636#step:22:428)

The K8sUpstreamNetConformance test seems like an important one to pass for this PR, since this function hasn't been touched in ~five years, but the changes are fairly trivial. @christarazi how would you like to go from here?