@christarazi 2023-06-20
@learnitall Have you checked the sysdumps / logs of Cilium to see if any of the errors come from the newly added error handling that this PR introduces? That is one easy way to check if this PR is relevant to the CI failures.