@learnitall 2023-06-21
After re-running and looking through the sysdump and logs for the failing net conformance test (here: https://github.com/cilium/cilium/actions/runs/5338165106/jobs/9675167448?pr=26238), I can't find any reference to the errors that I added. There are five new errors, and any one of them can be found with:

```
grep -E 'failed to (open|list|move)( old| new)? endpoint'
```